local_service/pyproject.toml
[project]
name = "friday-local-service"
version = "0.1.0"
description = "Friday local helper — wake detection, LiveKit token signing, local events API"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "pydantic>=2.9",
    "pydantic-settings>=2.6",
    "livekit-api>=0.8",
    "vosk==0.3.44",
    "sounddevice>=0.5",
    "numpy>=1.26",
    "python-dotenv>=1.0",
    "google-api-python-client>=2.140",
    "google-auth>=2.34",
    "google-auth-oauthlib>=1.2",
]

[tool.uv]
package = false
